refactor(parser): Refactor parser modules to enhance structure and readability
- Updated CStyle.hs to improve organization of parsing functions, added new declaration parsers (pScSig, pInfix, pForeign, pImport) and type parsers (pType, pTyApp, pAtomType, etc.).
- Removed unused functions from Core.hs to streamline the module.
- Enhanced Regular.hs by adding additional parsing functions for signatures, infix declarations, and literals, while restructuring existing parsers for clarity.
- Consolidated common parsing logic and improved comments for better understanding of the parsing process.
