+++ src/Malgo/Parser/CStyle.hs
@@ -8,15 +8,35 @@ import Effectful (Eff, IOE, type (:>))
 import Malgo.Features (Features)
 import Malgo.Module (ModuleName (..), Workspace, parseArtifactPath, pwdPath)
 import Malgo.Parser.Core
+  ( Parser,
+    captureRange,
+    decimal,
+    ident,
+    lexeme,
+    manyUnaryOp,
+    operator,
+    optional,
+    reserved,
+    reservedOperator,
+    skipPragma,
+    space,
+    symbol,
+  )
 import Malgo.Prelude
 import Malgo.Syntax
 import Malgo.Syntax.Extension
 import Text.Megaparsec hiding (optional, parse)
+import Text.Megaparsec.Char hiding (space)
+import Text.Megaparsec.Char.Lexer qualified as L
+
+-- * Entry Point
 
 -- | parseCStyle parses C-style apply syntax with parentheses and braces
 parseCStyle :: (IOE :> es, Workspace :> es, Features :> es) => FilePath -> TL.Text -> Eff es (Either (ParseErrorBundle TL.Text Void) (Module (Malgo Parse)))
 parseCStyle srcPath text = do
   runParserT (pModule <* eof) srcPath text
+
+-- * Module and Declaration Parsing
 
 -- | pModule parses a complete module using C-style syntax
 pModule :: (IOE :> es, Workspace :> es, Features :> es) => Parser es (Module (Malgo Parse))
@@ -46,16 +66,9 @@ pDecl = do
       pScDef
     ]
 
--- | pScDef parses value definitions with C-style expression syntax
-pScDef :: (Features :> es) => Parser es (Decl (Malgo Parse))
-pScDef = captureRange do
-  reserved "def"
-  name <- choice [ident, between (symbol "(") (symbol ")") operator]
-  reservedOperator "="
-  body <- pExpr
-  pure $ \range -> ScDef range name body
+-- * Declaration Parsers
 
--- | pCStyleDataDef parses C-style data definitions with parenthesized parameters
+-- | pDataDef parses C-style data definitions with parenthesized parameters
 -- > cStyleDataDef = "data" ident "(" (ident ("," ident)*)? ")" "=" constructor ("|" constructor)* ;
 -- > constructor = ident "(" (atomType ("," atomType)*)? ")" ;
 pDataDef :: Parser es (Decl (Malgo Parse))
@@ -72,7 +85,7 @@ pDataDef = captureRange do
       parameters <- pConstructorParams
       pure (,name,parameters)
 
--- | pCStyleTypeSynonym parses C-style type synonyms with parenthesized parameters
+-- | pTypeSynonym parses C-style type synonyms with parenthesized parameters
 -- > cStyleTypeSynonym = "type" ident "(" (ident ("," ident)*)? ")" "=" type ;
 pTypeSynonym :: Parser es (Decl (Malgo Parse))
 pTypeSynonym = captureRange do
@@ -82,6 +95,98 @@ pTypeSynonym = captureRange do
   reservedOperator "="
   ty <- pType
   pure $ \range -> TypeSynonym range name parameters ty
+
+-- | pScSig parses a value signature.
+--
+-- > scSig = "def" (ident | "(" operator")")":" type ;
+pScSig :: Parser es (Decl (Malgo Parse))
+pScSig = captureRange do
+  reserved "def"
+  name <- choice [ident, between (symbol "(") (symbol ")") operator]
+  reservedOperator ":"
+  ty <- pType
+  pure $ \range -> ScSig range name ty
+
+-- | pScDef parses value definitions with C-style expression syntax
+pScDef :: (Features :> es) => Parser es (Decl (Malgo Parse))
+pScDef = captureRange do
+  reserved "def"
+  name <- choice [ident, between (symbol "(") (symbol ")") operator]
+  reservedOperator "="
+  body <- pExpr
+  pure $ \range -> ScDef range name body
+
+-- | pInfix parses an infix declaration.
+--
+-- > infix = "infixl" decimal operator | "infixr" decimal operator | "infix" decimal operator ;
+pInfix :: Parser es (Decl (Malgo Parse))
+pInfix = captureRange do
+  choice
+    [ do
+        reserved "infixl"
+        precedence <- decimal
+        operator <- between (symbol "(") (symbol ")") operator
+        pure $ \range -> Infix range LeftA precedence operator,
+      do
+        reserved "infixr"
+        precedence <- decimal
+        operator <- between (symbol "(") (symbol ")") operator
+        pure $ \range -> Infix range RightA precedence operator,
+      do
+        reserved "infix"
+        precedence <- decimal
+        operator <- between (symbol "(") (symbol ")") operator
+        pure $ \range -> Infix range NeutralA precedence operator
+    ]
+
+-- | pForeign parses a foreign declaration.
+--
+-- > foreign = "foreign" "import" ident ":" type ;
+pForeign :: Parser es (Decl (Malgo Parse))
+pForeign = captureRange do
+  reserved "foreign"
+  reserved "import"
+  name <- ident
+  reservedOperator ":"
+  ty <- pType
+  pure $ \range -> Foreign range name ty
+
+-- | pImport parses an import declaration.
+--
+-- > import = "module" importList "=" "import" moduleName
+-- > importList = "{" ".." "}"
+-- >            | "{" importItem ("," importItem)* "}"
+-- >            | moduleName ;
+-- > importItem = ident | "(" operator ")" ;
+pImport :: (IOE :> es, Workspace :> es) => Parser es (Decl (Malgo Parse))
+pImport = captureRange do
+  reserved "module"
+  importList <- pImportList
+  reservedOperator "="
+  reserved "import"
+  moduleName <- pModuleName
+  pure $ \range -> Import range moduleName importList
+  where
+    pImportList = choice [try pAll, pSelected, pAs]
+    pAll = between (symbol "{") (symbol "}") do
+      symbol ".."
+      pure Malgo.Syntax.Extension.All
+    pSelected = between (symbol "{") (symbol "}") do
+      items <- sepBy pImportItem (symbol ",")
+      pure $ Selected items
+    pAs = As <$> pModuleName
+    pImportItem = choice [ident, between (symbol "(") (symbol ")") operator]
+    pModuleName = asIdent <|> asPath
+    asIdent = ModuleName <$> ident
+    asPath = do
+      path <- pStringLiteral
+      sourcePath <- (.sourceName) <$> getSourcePos
+      pwd <- lift pwdPath
+      sourcePath <- lift $ parseArtifactPath pwd sourcePath
+      path' <- lift $ parseArtifactPath sourcePath path
+      pure $ Artifact path'
+
+-- * C-Style Specific Helpers
 
 -- | pParameterList parses comma-separated parameters in parentheses
 pParameterList :: Parser es [(Range, Text)]
@@ -94,6 +199,86 @@ pParameterList = between (symbol "(") (symbol ")") (sepBy pParameter (symbol ","
 -- | pConstructorParams parses comma-separated types in parentheses
 pConstructorParams :: Parser es [Type (Malgo Parse)]
 pConstructorParams = between (symbol "(") (symbol ")") (sepBy pType (symbol ","))
+
+-- * Type Parsers
+
+-- | pType parses a type.
+--
+-- > type = tyapp "->" type
+-- >      | tyapp ;
+pType :: Parser es (Type (Malgo Parse))
+pType = makeExprParser pTyApp table
+  where
+    table =
+      [ [ InfixR do
+            start <- getSourcePos
+            reservedOperator "->"
+            TyArr . Range start <$> getSourcePos
+        ]
+      ]
+
+-- | pTyApp parses a type application.
+--
+-- > tyapp = atomType atomType* ;
+pTyApp :: Parser es (Type (Malgo Parse))
+pTyApp = captureRange do
+  ty <- pAtomType
+  tys <- many pAtomType
+  pure $ \range -> case tys of
+    [] -> ty
+    _ -> TyApp range ty tys
+
+-- | pAtomType parses an atomic type.
+--
+-- > atomType = tyVar
+-- >          | tyTuple
+-- >          | tyRecord
+-- >          | tyBlock
+pAtomType :: Parser es (Type (Malgo Parse))
+pAtomType = choice [pTyVar, pTyTuple, try pTyRecord, pTyBlock]
+
+-- | pTyVar parses a type variable.
+--
+-- > tyVar = ident ;
+pTyVar :: Parser es (Type (Malgo Parse))
+pTyVar = captureRange do
+  name <- ident
+  pure $ \range -> TyVar range name
+
+-- | pTyTuple parses a tuple type or parenthesized type.
+--
+-- > tyTuple = "(" type ("," type)* ")"
+-- >         | "(" ")" ;
+pTyTuple :: Parser es (Type (Malgo Parse))
+pTyTuple = captureRange do
+  tys <- between (symbol "(") (symbol ")") (sepBy pType (symbol ","))
+  pure $ \range -> case tys of
+    [ty] -> ty
+    _ -> TyTuple range tys
+
+-- | pTyRecord parses a record type.
+--
+-- > tyRecord = "{" ident "=" type ("," ident "=" type)* "}" ;
+pTyRecord :: Parser es (Type (Malgo Parse))
+pTyRecord = captureRange do
+  fields <- between (symbol "{") (symbol "}") $ sepEndBy1 pField (symbol ",")
+  pure $ \range -> TyRecord range fields
+  where
+    pField = do
+      field <- ident
+      reservedOperator ":"
+      value <- pType
+      pure (field, value)
+
+-- | pTyBlock parses a block type.
+--
+-- > tyBlock = "{" type "}" ;
+pTyBlock :: Parser es (Type (Malgo Parse))
+pTyBlock = captureRange do
+  ty <- between (symbol "{") (symbol "}") pType
+  pure $ \range -> TyBlock range ty
+
+-- * Expression Parsers
 
 -- | pExpr parses expressions with C-style syntax
 pExpr :: (Features :> es) => Parser es (Expr (Malgo Parse))
@@ -184,47 +369,18 @@ pFn = captureRange do
   clauses <- between (symbol "{") (symbol "}") $ sepEndBy1 pClause (symbol ",")
   pure $ \range -> Fn range $ NonEmpty.fromList clauses
 
--- | pClause parses C-style clauses with optional parentheses
--- > clause = "(" pattern ("," pattern)* ")" "->" stmts
--- >        | pattern+ "->" stmts
-pClause :: (Features :> es) => Parser es (Clause (Malgo Parse))
-pClause = captureRange do
-  patterns <-
-    try (between (symbol "(") (symbol ")") (sepEndBy pAtomPat (symbol ",")) <* reservedOperator "->")
-      <|> try (some pAtomPat <* reservedOperator "->")
-      <|> pure []
-  body <- pStmts
-  pure $ \range -> Clause range patterns body
+-- | pList parses list literals using C-style syntax
+pList :: (Features :> es) => Parser es (Expr (Malgo Parse))
+pList = between (symbol "[") (symbol "]") do
+  captureRange do
+    elements <- sepEndBy pExpr (symbol ",")
+    pure $ \range -> List range elements
 
--- | pAtomPat parses atomic patterns with C-style syntax
-pAtomPat :: (Features :> es) => Parser es (Pat (Malgo Parse))
-pAtomPat =
-  choice
-    [ pVarP,
-      pLiteralP,
-      try pTupleP,
-      pRecordP,
-      pListP
-    ]
+-- | pSeq parses parenthesized sequences using C-style syntax
+pSeq :: (Features :> es) => Parser es (Expr (Malgo Parse))
+pSeq = between (symbol "(") (symbol ")") pStmts
 
--- | pTupleP parses C-style tuple patterns with braces
-pTupleP :: (Features :> es) => Parser es (Pat (Malgo Parse))
-pTupleP = captureRange do
-  patterns <- between (symbol "{") (symbol "}") (sepBy pPat (symbol ","))
-  case patterns of
-    [_] -> fail "c-style tuple must have at least two patterns or be empty"
-    _ -> pure $ \range -> TupleP range patterns
-
--- | pPat parses patterns with C-style syntax
-pPat :: (Features :> es) => Parser es (Pat (Malgo Parse))
-pPat = try pConP <|> pAtomPat
-
--- | pConP parses constructor patterns with C-style syntax
-pConP :: (Features :> es) => Parser es (Pat (Malgo Parse))
-pConP = captureRange do
-  constructor <- ident
-  patterns <- some pAtomPat
-  pure $ \range -> ConP range constructor patterns
+-- * Statement Parsers
 
 -- | pStmts parses statements using C-style syntax
 pStmts :: (Features :> es) => Parser es (Expr (Malgo Parse))
@@ -266,16 +422,37 @@ pNoBind = captureRange do
   body <- pExpr
   pure $ \range -> NoBind range body
 
--- | pSeq parses parenthesized sequences using C-style syntax
-pSeq :: (Features :> es) => Parser es (Expr (Malgo Parse))
-pSeq = between (symbol "(") (symbol ")") pStmts
+-- * Pattern Parsers
 
--- | pList parses list literals using C-style syntax
-pList :: (Features :> es) => Parser es (Expr (Malgo Parse))
-pList = between (symbol "[") (symbol "]") do
-  captureRange do
-    elements <- sepEndBy pExpr (symbol ",")
-    pure $ \range -> List range elements
+-- | pPat parses patterns with C-style syntax
+pPat :: (Features :> es) => Parser es (Pat (Malgo Parse))
+pPat = try pConP <|> pAtomPat
+
+-- | pAtomPat parses atomic patterns with C-style syntax
+pAtomPat :: (Features :> es) => Parser es (Pat (Malgo Parse))
+pAtomPat =
+  choice
+    [ pVarP,
+      pLiteralP,
+      try pTupleP,
+      pRecordP,
+      pListP
+    ]
+
+-- | pConP parses constructor patterns with C-style syntax
+pConP :: (Features :> es) => Parser es (Pat (Malgo Parse))
+pConP = captureRange do
+  constructor <- ident
+  patterns <- some pAtomPat
+  pure $ \range -> ConP range constructor patterns
+
+-- | pTupleP parses C-style tuple patterns with braces
+pTupleP :: (Features :> es) => Parser es (Pat (Malgo Parse))
+pTupleP = captureRange do
+  patterns <- between (symbol "{") (symbol "}") (sepBy pPat (symbol ","))
+  case patterns of
+    [_] -> fail "c-style tuple must have at least two patterns or be empty"
+    _ -> pure $ \range -> TupleP range patterns
 
 -- | pRecordP parses record patterns using C-style syntax
 pRecordP :: (Features :> es) => Parser es (Pat (Malgo Parse))
@@ -294,3 +471,107 @@ pListP :: (Features :> es) => Parser es (Pat (Malgo Parse))
 pListP = captureRange do
   pats <- between (symbol "[") (symbol "]") $ sepEndBy pPat (symbol ",")
   pure $ \range -> ListP range pats
+
+-- | pVarP parses a variable pattern.
+--
+-- > varPat = ident ;
+pVarP :: Parser es (Pat (Malgo Parse))
+pVarP = captureRange do
+  name <- ident
+  pure $ \range -> VarP range name
+
+-- * Literal Parsers
+
+-- | pLiteral parses a literal.
+--
+-- > literal = boxed ["#"] ;
+-- > boxed = double | float | int | long | char | string ;
+pLiteral :: Parser es (Expr (Malgo Parse))
+pLiteral = captureRange do
+  boxed <- pBoxed
+  sharp <- optional (symbol "#")
+  pure $ \range -> case sharp of
+    Just _ -> Unboxed range $ coerce boxed
+    Nothing -> Boxed range boxed
+
+-- | pLiteralP parses a literal pattern.
+pLiteralP :: Parser es (Pat (Malgo Parse))
+pLiteralP = captureRange do
+  boxed <- pBoxed
+  sharp <- optional (symbol "#")
+  pure $ \range -> case sharp of
+    Just _ -> UnboxedP range $ coerce boxed
+    Nothing -> BoxedP range boxed
+
+pBoxed :: Parser es (Literal Boxed)
+pBoxed = choice [try pReal, pInt, pChar, pString]
+
+-- | pReal parses a real number.
+--
+-- > real = double | float ;
+-- > double = FLOAT ;
+-- > float = FLOAT "f" | FLOAT "F" ;
+pReal :: Parser es (Literal Boxed)
+pReal = lexeme do
+  f <- L.float
+  tail <- optional (char 'f' <|> char 'F')
+  case tail of
+    Just _ -> pure $ Float (realToFrac f)
+    Nothing -> pure $ Double f
+
+-- | pInt parses an integer.
+--
+-- > int = int32| int64 ;
+-- > int32 = DECIMAL ;
+-- > int64 = DECIMAL "l" | DECIMAL "L" ;
+pInt :: Parser es (Literal Boxed)
+pInt = lexeme do
+  i <- L.decimal
+  tail <- optional (char 'l' <|> char 'L')
+  case tail of
+    Just _ -> pure $ Int64 i
+    Nothing -> pure $ Int32 (fromIntegral i)
+
+-- | pChar parses a character.
+--
+-- > char = "'" CHAR "'" ;
+pChar :: Parser es (Literal Boxed)
+pChar = lexeme do
+  void $ char '\''
+  c <- L.charLiteral
+  void $ char '\''
+  pure $ Char c
+
+-- | pString parses a string.
+--
+-- > string = "\"" CHAR* "\"" ;
+pString :: Parser es (Literal Boxed)
+pString = String <$> pStringLiteral
+
+pStringLiteral :: (ConvertibleStrings String a) => Parser es a
+pStringLiteral = lexeme do
+  void $ char '"'
+  str <- manyTill L.charLiteral (char '"')
+  pure $ convertString str
+
+-- * Utility Functions
+
+-- | pVariable parses a variable.
+--
+-- > variable = ident ;
+pVariable :: Parser es (Expr (Malgo Parse))
+pVariable = captureRange do
+  name <- ident
+  pure $ \range -> Var range name
+
+-- | pClause parses C-style clauses with optional parentheses
+-- > clause = "(" pattern ("," pattern)* ")" "->" stmts
+-- >        | pattern+ "->" stmts
+pClause :: (Features :> es) => Parser es (Clause (Malgo Parse))
+pClause = captureRange do
+  patterns <-
+    try (between (symbol "(") (symbol ")") (sepEndBy pAtomPat (symbol ",")) <* reservedOperator "->")
+      <|> try (some pAtomPat <* reservedOperator "->")
+      <|> pure []
+  body <- pStmts
+  pure $ \range -> Clause range patterns body
